src/patient_service/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.3.0
redis==4.6.0
sqlalchemy==2.0.19
pytest==8.3.2
httpx==0.24.0
google-cloud-pubsub==2.23.0
python-dotenv==1.0.1
